Addresses comments and refactored code.
